// src/app/types/content.ts
export type LandingInfoFields = {
  welcome?: string;
  landingSub?: string;
  landingContent?: string;
  subTitle?: string;
  features?: string[];
};

export type FeaturesFields = {
  subTitle: string;
  features: string[];
};

export type AboutFields = {
  title: string;
  description: string;
  complementaryDescription: string;
  subTitle: string;
  subDescription: string;
  listNames: string[];
};